Improved folder structure for tests and moved store.
